refactor: Refactored the secrets.py module to remove the          env and vault secret manager code and refacotred them into their own modules in a new secretmanager package. Also refacotored the tests
