+++ tests/secretmanagers/test_vault.py
@@ -2,72 +2,15 @@ import os
 from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
-from fixtures import config
+from tests.fixtures import config
 from hvac.exceptions import InvalidRequest
 
 from chaoslib.exceptions import InvalidExperiment
-from chaoslib.secret import create_vault_client, load_secrets
+from chaoslib.secretmanagers.vault import create_vault_client
+from chaoslib.secret import load_secrets
 
 
-@patch.dict(os.environ, {"KUBE_API_URL": "http://1.2.3.4"})
-def test_should_load_environment():
-    secrets = load_secrets(
-        {
-            "kubernetes": {
-                "api_server_url": {"type": "env", "key": "KUBE_API_URL"}
-            }
-        },
-        config.EmptyConfig,
-    )
-    assert secrets["kubernetes"]["api_server_url"] == "http://1.2.3.4"
-
-
-@patch.dict(os.environ, {"KUBE_API_URL": "http://1.2.3.4"})
-def test_should_load_nested_environment():
-    secrets = load_secrets(
-        {
-            "kubernetes": {
-                "env1": {
-                    "username": "jane",
-                    "address": {"host": "whatever", "port": 8090},
-                    "api_server_url": {"type": "env", "key": "KUBE_API_URL"},
-                }
-            }
-        },
-        config.EmptyConfig,
-    )
-    assert secrets["kubernetes"]["env1"]["username"] == "jane"
-    assert secrets["kubernetes"]["env1"]["address"]["host"] == "whatever"
-    assert secrets["kubernetes"]["env1"]["address"]["port"] == 8090
-    assert secrets["kubernetes"]["env1"]["api_server_url"] == "http://1.2.3.4"
-
-
-def test_should_load_inline():
-    secrets = load_secrets(
-        {"kubernetes": {"api_server_url": "http://1.2.3.4"}}, config.EmptyConfig
-    )
-    assert secrets["kubernetes"]["api_server_url"] == "http://1.2.3.4"
-
-
-@patch.dict(os.environ, {"KUBE_API_URL": "http://1.2.3.4"})
-def test_should_merge_properly():
-    secrets = load_secrets(
-        {
-            "kubernetes": {
-                "username": "jane",
-                "address": {"host": "whatever", "port": 8090},
-                "api_server_url": {"type": "env", "key": "KUBE_API_URL"},
-            }
-        },
-        config.EmptyConfig,
-    )
-    assert secrets["kubernetes"]["username"] == "jane"
-    assert secrets["kubernetes"]["address"]["host"] == "whatever"
-    assert secrets["kubernetes"]["address"]["port"] == 8090
-    assert secrets["kubernetes"]["api_server_url"] == "http://1.2.3.4"
-
-
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_should_auth_with_approle(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -89,7 +32,7 @@ def test_should_auth_with_approle(hvac):
     )
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_should_catch_approle_invalid_secret_id_abort_the_run(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -105,7 +48,7 @@ def test_should_catch_approle_invalid_secret_id_abort_the_run(hvac):
         create_vault_client(config)
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_should_auth_with_token(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -122,7 +65,7 @@ def test_should_auth_with_token(hvac):
     fake_client.auth_approle.assert_not_called()
 
 
-@patch("chaoslib.secret.hvac", autospec=True)
+@patch("chaoslib.secretmanagers.vault.hvac", autospec=True)
 def test_should_auth_with_service_account(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -134,7 +77,7 @@ def test_should_auth_with_service_account(hvac):
     fake_client = MagicMock()
     hvac.Client.return_value = fake_client
 
-    with patch("chaoslib.secret.open", mock_open(read_data="fake_sa_token")):
+    with patch("chaoslib.secretmanagers.vault.open", mock_open(read_data="fake_sa_token")):
         vault_client = create_vault_client(config)
         vault_client.auth_approle.assert_not_called()
         vault_client.auth_kubernetes.assert_called_with(
@@ -145,7 +88,7 @@ def test_should_auth_with_service_account(hvac):
         )
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_should_catch_service_account_invalid_abort_the_run(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -161,7 +104,7 @@ def test_should_catch_service_account_invalid_abort_the_run(hvac):
         create_vault_client(config)
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_read_secrets_from_vault_with_kv_version_1(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -207,7 +150,7 @@ def test_read_secrets_from_vault_with_kv_version_1(hvac):
     assert secrets["k8s"]["a-secret"] == "bar"
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_read_secrets_from_vault_with_kv_version_2(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -282,7 +225,7 @@ def test_should_override_load_inline_with_var():
     assert secrets["kubernetes"]["api_server_url"] == "http://elsewhere"
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_vault_add_subkeys(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -316,7 +259,7 @@ def test_vault_add_subkeys(hvac):
     assert secrets["myapp"]["token"]["baz"] == "hello"
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_vault_replace_entire_declare(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
@@ -357,7 +300,7 @@ def test_vault_replace_entire_declare(hvac):
     assert secrets["myapp"]["token"] == "bar"
 
 
-@patch("chaoslib.secret.hvac")
+@patch("chaoslib.secretmanagers.vault.hvac")
 def test_override_vault_with_vars(hvac):
     config = {
         "vault_addr": "http://someaddr.com",
